put in cast to ldexp() call

+++ src/common/readoct.c
@@ -173,7 +173,7 @@ getflt()			/* get a floating point number */
 	double  d;
 
 	d = (double)getint(4)/0x7fffffff;
-	return(ldexp(d, getint(1)));
+	return(ldexp(d, (int)getint(1)));
 }
